+++ crates/autopilot/src/infra/mod.rs
@@ -1,5 +1,3 @@
 pub mod blockchain;
-pub mod orderbook;
 pub mod persistence;
-
-pub use orderbook::Orderbook;
+pub mod shadow;
