+++ crates/autopilot/src/infra/shadow/mod.rs
@@ -1,0 +1,3 @@
+pub mod orderbook;
+
+pub use orderbook::Orderbook;
