+++ crates/autopilot/src/infra/shadow/orderbook/mod.rs
@@ -1,9 +1,21 @@
-use crate::{
-    domain,
-    infra::{self, persistence::auction::dto},
+//! A client to the CoW Protocol public API.
+
+use {
+    crate::{domain, infra::persistence::auction::dto},
+    reqwest::Url,
 };
 
-impl infra::Orderbook {
+pub struct Orderbook {
+    client: reqwest::Client,
+    url: Url,
+}
+
+impl Orderbook {
+    /// Creates a new CoW Protocol client.
+    pub fn new(client: reqwest::Client, url: Url) -> Self {
+        Self { client, url }
+    }
+
     /// Retrieves the current auction.
     pub async fn auction(&self) -> reqwest::Result<domain::AuctionWithId> {
         self.client
